+++ src/system/rndobj/Utl.h
@@ -43,15 +43,4 @@ DataNode GetNormalMapTextures(ObjectDir*);
 DataNode GetRenderTextures(ObjectDir*);
 DataNode GetRenderTexturesNoZ(ObjectDir*);
 
-// idk where else to put this
-template <class T1, class T2, class T3>
-void VectorRemove(std::vector<T1,T2>& vec, const T3& obj){
-    for(std::vector<T1,T2>::iterator it = vec.begin(); it != vec.end(); ++it){
-        if(*it == obj){
-            vec.erase(it);
-            return;
-        }
-    }
-}
-
 #endif // RNDOBJ_UTL_H
